cloud_functions/is_new_buyer/requirements.txt
bloom-filter==1.3
google-cloud-core==1.0.2
google-api-python-client==1.7.9
google-cloud-datastore==1.8.0